requirements.txt
# Google Gemini AI
google-genai>=1.0.0

# Mistral AI (OCR)
mistralai>=1.0.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
lxml>=5.0.0

# PDF Parsing (fallback)
pdfplumber>=0.10.0

# Data Processing
pydantic>=2.5.0
pandas>=2.1.0

# CLI & Utilities
rich>=13.7.0
python-dotenv>=1.0.0

# Async Support
aiohttp>=3.9.0
asyncio-throttle>=1.0.0
